i9500: move routing structures to separate header file
